YYYYEEEEEEEEEEEEEEEEEEe I finally debugged the hash!!!!! The problem was due to an error in usage o FileTreeHashBuilder during unpacking.

+++ src/main/java/net/szum123321/textile_backup/core/digest/BalticHash.java
@@ -38,8 +38,8 @@ public class BalticHash implements Hash {
     protected final ByteBuffer buffer = ByteBuffer.wrap(_byte_buffer).order(ByteOrder.LITTLE_ENDIAN);
     protected long hashed_data_length = 0;
 
-    public void update(byte b) {
-        buffer.put(b);
+    public void update(int b) {
+        buffer.put((byte)b);
         hashed_data_length += 1;
         if (buffer.position() >= buffer_limit) round();
     }
